4.21*: bloglist expansion, step9 and many fix!

+++ part4/Bloglist/controllers/blogs.js
@@ -8,18 +8,10 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs);
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    
-    if(!request.get('authorization')) {
-        return response.status(401).json({ error: 'no token' })
-    }
+
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
 
     const user = await User.findById(decodedToken.id)
 
@@ -59,12 +51,16 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
+blogsRouter.delete('/:id', async (request, response) => {
+
+    const blog = await Blog.findById(request.params.id);
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
         await Blog.findByIdAndRemove(request.params.id);
         response.status(204).end();
-    } catch (error) {
-        next(error)
+    } else {
+        response.status(401).json({ error: "user isnt creator therefore can't delete blog" })
     }
 })
